(dev-VER 0.5.2 BETA-1.0)
Moving to the Visual Studio Debugger using VMware (because debugging this Malware on my local Machine would be fucking stupid).

I finally want to debug my Malware with a proper Debugger (MessageBox debugging worked fine for the start but now it isn't enough anymore.
I need other/better way's to debug this thing, like the memory viewer from VS to debug buffers and pointers, Breakpoints, etc.)

In order to make this as easy for me as possible without ignoring safety.
For this I use the batchscripts to easily and safely copy the Solution between VMrepo, VMshare and Local.
The Directories are hardcoded in these scripts so if you want to use them you'll need to modifie them

Im also starting to implement a console version for Debugmode